fixed a bunch of warnings
git-svn-id: https://bwapi.googlecode.com/svn/trunk@2286 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPILIB/Source/PlayerType.cpp
@@ -49,7 +49,7 @@ namespace BWAPI
       for(std::set<PlayerType>::iterator i = playerTypeSet.begin(); i != playerTypeSet.end(); i++)
       {
         std::string name=(*i).getName();
-        for(int j=0;j<name.length();j++)
+        for(int j=0;j<(int)name.length();j++)
         {
           if (name[j]==' ') name[j]='_';
           if (name[j]>='a' && name[j]<='z') name[j]+='A'-'a';
@@ -105,7 +105,7 @@ namespace BWAPI
   }
   PlayerType PlayerTypes::getPlayerType(std::string& name)
   {
-    for(int j=0;j<name.length();j++)
+    for(int j=0;j<(int)name.length();j++)
     {
       if (name[j]==' ') name[j]='_';
       if (name[j]>='a' && name[j]<='z') name[j]+='A'-'a';
