+++ bwapi/BWAPILIB/Source/UnitType.cpp
@@ -605,7 +605,7 @@ namespace BWAPI
       for(std::set<UnitType>::iterator i = unitTypeSet.begin(); i != unitTypeSet.end(); i++)
       {
         std::string name=(*i).getName();
-        for(int j=0;j<name.length();j++)
+        for(int j=0;j<(int)name.length();j++)
         {
           if (name[j]==' ') name[j]='_';
           if (name[j]>='a' && name[j]<='z') name[j]+='A'-'a';
@@ -922,7 +922,7 @@ namespace BWAPI
 
   UnitType UnitTypes::getUnitType(std::string& name)
   {
-    for(int j=0;j<name.length();j++)
+    for(int j=0;j<(int)name.length();j++)
     {
       if (name[j]==' ') name[j]='_';
       if (name[j]>='a' && name[j]<='z') name[j]+='A'-'a';
